Moved options into options.go
